[avr] Add more documentation and clean up some code

+++ avr/src/Blink.cpp
@@ -2,7 +2,7 @@
 
 #include <Arduino.h>
 
-Blink::Blink(uint8_t pin, int delay) : m_pin(pin), m_enabled(false), m_delay(delay)
+Blink::Blink(uint8_t pin, unsigned long delay) : m_pin(pin), m_enabled(false), m_delay(delay)
 {
 	pinMode(pin, OUTPUT);
 	digitalWrite(m_pin, LOW);
